MonitoringStackLocal/samples/pyproject.toml
[project]
name = "samples"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "<4.0,>=3.12"
dependencies = [
    "extensions-opentelemetry (>=0.2.0,<0.3.0)"
]


[[tool.poetry.source]]
name = "local-pypi"
url = "http://pypi.mw.local"

[[tool.poetry.source]]
name = "PyPI"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
